SDK-253 file name standard refactoring
