+++ .gitignore
@@ -25,3 +25,5 @@ uploads
 
 #visual studio code
 .vscode
+
+coverage
