+++ modules/app/client/app-routing.module.ts
@@ -1,9 +1,7 @@
 import { NgModule }				 from '@angular/core';
 import { RouterModule, Routes }  from '@angular/router';
 
-import { CoreMenuComponent }	 from '../../core/client/components/core.component.menu.client';
 import { PageNotFoundComponent } from './components/not-found.component';
-
 
 /* this defines the global app routes */
 const appRoutes: Routes = [
