+++ modules/auth/client/auth.module.ts
@@ -3,7 +3,7 @@ import { NgModule }             from '@angular/core';
 import { BrowserModule }        from '@angular/platform-browser';
 
 /*  Modules */
-import { AuthRoutingModule }    from './config/auth.route';
+import { AuthRoutingModule }    from './config/auth-routing.module';
 
 /*  Components */
