+++ modules/core/client/core.module.ts
@@ -3,7 +3,7 @@ import { BrowserModule }  from '@angular/platform-browser';
 import { HttpModule }     from '@angular/http';
 import { NgbModule }      from '@ng-bootstrap/ng-bootstrap';
 import { RouterModule, Routes }     from '@angular/router';
-import { CoreMenuComponent }  from './components/core.component.menu.client';
+import { CoreMenuComponent }  from './components/core.component';
 import { RoleModule } from '../../roles/client/roles.module';
 import { AppModule }      from '../../app/client/app.module';
 import { AppRoutingModule }      from '../../app/client/app-routing.module';
