+++ modules/core/client/views/core.view.html
