+++ modules/users/client/components/admin-users.component.ts
@@ -11,10 +11,10 @@ import { NgbModule, NgbModalModule, NgbModal, NgbModalOptions, NgbActiveModal, N
 import { ViewContainerRef } from '@angular/core';
 import { Subject } from 'rxjs/Subject';
 /*  Angular2 Models */
-import { User }             from './../models/user.model.client';
+import { User }             from './../models/user.model';
 
 /*  Angular2 Services */
-import { AuthService }      from './../../../auth/client/auth.service.client';
+import { AuthService }      from './../../../auth/client/services/auth.service';
 import { UserService }      from '../services/user.service';
 import { AdminUserForm }   from './admin-user-form.component';
