+++ modules/users/client/components/profile.component.ts
@@ -4,10 +4,10 @@ import { BrowserModule }    from '@angular/platform-browser';
 import { Router }           from '@angular/router';
 
 /* Angular2 Models */
-import { User }             from '../models/user.model.client';
+import { User }             from '../models/user.model';
 
 /* Angular2 Services */
-import { AuthService }      from '../../../auth/client/auth.service.client';
+import { AuthService }      from '../../../auth/client/services/auth.service';
 import { UserService }      from '../services/user.service';
