+++ modules/users/client/components/signin.component.spec.ts
@@ -15,8 +15,8 @@ import { SigninComponent }                  from './signin.component';
 
 import { UsersRoutingModule } from '../config/user-routing.module';
 import { UserService } from '../services/user.service';
-import { AuthService } from '../../../auth/client/auth.service.client';
-import { User } from '../models/user.model.client';
+import { AuthService } from '../../../auth/client/services/auth.service';
+import { User } from '../models/user.model';
 
 import { UsersComponent }       from './users.component';
 import {
